Re-written this whole paragraph to postpone the ambiguity to chapter 2.
p4., l.10: you claim the plasma in the photosphere is tied to the magnetic field. But there not only the ionization degree is only 0.01

, but also the density is so high that for ions the collision frequency is four orders of magnitude larger than the gyro frequency, I.e. the ions collide 10.000 times in one turn around the magnetic field line. SO they are ‘unmagnetized’!?;

p.4, l.-6: the statement on magnetic reconnection contradicts the previous statement on perfectly conducting plasma…?;

p.6, l.4: plasma and magnetic field are locked together in MHD simulations of the motions in the photosphere, but this is by definition (MHD). Can we be sure it is also true on the real Sun, given the remarks above?