Fixed by explaining notation.
p.10, Eq.(2.16): to get this simple form of decoupled equations you have to consider a coordinate system chosen such that the (x.z)-plane is the plane determined by the wave vector k and the magnetic field B. If ky is not equal to zero, the equations are coupled!

Fixed by explaining notation.

p.10, Eq.(2.16): also, you seem to use cylindrical coordinates? what is the meaning of phi then, as I though phi is the perpendicular direction!!

This is fixed by correctly explaining the notation convention. It is in fact in Cartesian coordinates.

there should be curvature term in the equation (2.16) when you use cylindrical coordinates!

Changed all notation up to this point to use this notation.

p.15, section 2.1.3: why this sudden change of notation with tildes on the perturbed ? = B1…? and v = v1? quantities? I guess you mean Bb = B0 and 𝐁𝐁