Notation fixed.
p.34: I do not understand this driver. The velocity is a VECTOR field but Eq.(4.1) seems to define a scalar field!? What are the components of the velocity in the different cases? In your slides (short presentation at the viva) you actually had vectors, but this should be also corrected in the thesis text.

Fixed.

p.34-35: Actually, also in the following equations (4.2)-(4.4) your notations is WRONG as these are the components of the vector field F, so F(x) should be Fx and F(y) = Fy.