N/A
p.38 and Fig.4.3b: vertical is parallel (vertical field), right? On page 15 you claimed the fast mode perturbs mostly the parallel component of the velocity. So this is the one you drive directly in Fig.4.3b (vertical = parallel to B))? So why is the Alfvén wave dominant in the response? You write in the text that the phi component is ‘substantially weaker than the two other components’, but this is clearly not true!!? This turns out to be a scaling problem, the color scale is different in the different plots!!!

Added small clarification. Again the scaling on the plots are important.

p.38 and Fig.4.3c-e: how do you explain the asymmetry in these cases (much more asymmetric response than the others)? Also, you claim in the text that for these three torsional driving cases the vast majority of the perturbation is in the phi component. This is not true as it seems to me that the perpendicular component is dominant!? Or is this just an impression from the larger colored areas in the plots?