Explained in text by referencing the vertical driver in ch4.
TODO Fig.5.3e, p.56: Are you sure the panels have not been interchanged. For B_L = 1.5 the velocity is almost radial (see Fig.5.2.c), so one would expect most energy in the perpendicular component (which is the radial component here , is it not?), and not in the phi component which is almost perpendicular to that, and so should receive almost no energy…!?

Explained by above.

P.61, Fig.5.5: Again, it is strange that her for BL = 1.5 most energy seems to go to the fast (parallel) component!? Even though you are driving horizontally in this case, or not?