Discussion point.
The effect of the driving motions on the excitation of MHD waves actually depends on the oscillation spectrum of the MBP structure you are modeling. Do you have any idea of the eigenvalue spectrum of this configuration?

Discussion point.

Do you have any idea on how it could be computed, i.e. how you could formulate the problem as an eigenvalue problem instead of an initial value problem?

The scales on the plots are slightly different, but the results are the same.

p.80, Fig.6.4: I do not understand the contrast of this figure with Fig.5.5. In fact, the case P = 180 s is the same as in Fig.5.5, right? But the results seem different here? Now the phi component is by far the most energetic while it used to be comparable to the parallel component in Fig.5.5. What happened?