This is attributed to the intertia of the system after it has been accelerated from the stationary initial conditions.
p.70, Fig.6.2: It had occurred to me earlier but in these figures it is more apparent that in the phi and perpendicular components of the velocity you do not get an oscillatory response at all!? As a matter of fact, the phi component does not become positive and the perpendicular component does not become negative. So in these components the response is more like a series of pulses, rather than an oscillatory wave. Only in the parallel component you seem to get oscillations!?