Added comments about limitation of the model. Discussion of the limitations of the model when applied to the simulations is in the results sections of the proceeding chapters.
Actually, in the real Sun the situation is far more complex. When inhomogeneity of the plasma is taken into account, the MHD wave spectrum includes continuous parts with singular wave mode solutions. Did this not come up and cause problems in the simulations? The waves should phase-mix as they propagate along the magnetic field lines, especially in the case of the circular driver, when you impose a phi-component only, i.e. directly drive an Alfvén wave…