Added clarification about derivation in Leroy 1985.
P.15, Eq.(2.66): this I a second order equation/quantity, is this OK since you linearized the equations before!? Linear is first order… Actually, why do you consider only the second order form of this wave flux, instead of the full nonlinear form?

This is discussed when talking about why this equation for the linear MHD wave flux was chosen.

And why do you not simply consider the Poynting flux???