Added in table 2.1
p.7, middle: can you give an idea on the typical time and length scales in the photosphere/chromosphere/corona and what you mean by ‘sufficiently large’ then?

Added. Also added a discussion regarding ionisation ratio and effects on MHD waves.

You should include a quantitative check of the satisfaction of the assumptions to show the MHD model can indeed be used in this situation, i.e. all the assumptions that are made to derive the MHD equations are satisfied!

done.

p.7, l.-6: are the relative velocities of positive and negative charges small, given the statements above and the large discrepancy between the mass of electrons and ions?